requirements.txt
Package      Version
------------ -------
pip          22.0.4
pycryptodome 3.14.1
setuptools   60.6.0
wheel        0.37.1
